update: tests/hakikisha.ts
Previously, all edge cases apart from the first two were failing. Using
a similar wallet (manufacturerKp) within setupManufacturerAndRetailer()
was causing an 'already in use' error. Re-registering manufacturer and
retailer accounts within this function was the source of the error.
Currently, the function has been changed to generate a new Keypair for
the manufacturer and retailer, fund them and also generate PDAs for
them.